Using Pocket's OAuth for API calls.
This means that the app will work for anyone with a Pocket account intead
of just me.

I now show the username and have a way to logout.

Also get a page at a time, and only retrieve more items once the current
page is empty.

Updated .env.example with the proper values (got rid of three)