Using redis sessions for tokens.
Moved logic to store and retrieve tokens (request and access) into the
server for security. Now the only thing stored in local storage is the
username and the list of items to serve as a cache.

Added necessary routes for new auth.

Enhanced hot pocket image to be clearer.

Removed all mention of access and request tokens from Vue component.

Moved dev dependencies into correct location in package.json.