Cleaner cache-busting solution.
Instead of having to add/delete the generated CSS and JS bundles to Git,
gulp now writes the timestamp to a .timestamp.php file that is not version
controlled. index.php includes the .timestamp.php file and uses it to
reference the CSS and JS files.

This means that when deploying, in addition to pulling the new changes
down, gulp needs to be run. But when gulp is run, there will be no local
changes on the server.

Also improved error handling in the client when the session has expired
and we need to logout.