Upgrading to latest version of guzzle
Note it doesn't resolve this vulnerability because it's not using the
latest symfony/http-foundation package

https://github.com/jaden/HotPocket/network/alert/composer.lock/symfony%2Fhttp-foundation/open

There are also notes about Carbon upgrades and if I want to upgrade to
the latest Lumen version I have to stop using sessions and upgrade to
Laravel instead. Dependencies are great until you have to deal with this
sort of stuff. I'm just going to leave things as they are for now and
hope that guzzle puts out a new release to use the latest
http-foundation package.